Django RESTful

+++ Web/04_django/Open_API/DJANGO_BLOG/articles/views.py
@@ -7,28 +7,31 @@ def index(request):
     # articles = Article.objects.all()[::-1]
     return render(request, 'articles/index.html', {'articles':articles})
 
-def new(request):
-    return render(request, 'articles/new.html')
+# def new(request):
+#     return render(request, 'articles/new.html')
 
 def create(request):
-    title = request.POST.get('title')
-    content = request.POST.get('content')
+    if request.method == "POST":
+        title = request.POST.get('title')
+        content = request.POST.get('content')
 
-    article = Article(title=title, content=content)
-    article.save()
+        article = Article(title=title, content=content)
+        article.save()
 
-    # 2.
-    # article = Article()
-    # article.title = title
-    # article.content = content
-    # article.save()
+        # 2.
+        # article = Article()
+        # article.title = title
+        # article.content = content
+        # article.save()
 
-    # 3.
-    # Article.objects.create(title=title, content=content)
+        # 3.
+        # Article.objects.create(title=title, content=content)
 
-    # return render(request, 'articles/create.html')
-    # return redirect('/articles/')    # import redirect
-    return redirect(f'/articles/{article.id}/')
+        # return render(request, 'articles/create.html')
+        # return redirect('/articles/')    # import redirect
+        return redirect('articles:detail', article.id)
+    else:
+        return render(request, 'articles/new.html')
 
 def detail(request, id):
     article = Article.objects.get(pk=id)
@@ -36,16 +39,23 @@ def detail(request, id):
 
 def delete(request, id):
     article = Article.objects.get(pk=id)
-    article.delete()
-    return redirect('/articles/')
+    if request.method == "POST":
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return redirect('articles:detail', article.id)
+    
 
-def edit(request, id):
-    article = Article.objects.get(pk=id)
-    return render(request, 'articles/edit.html', {'article':article})
+# def edit(request, id):
+#     article = Article.objects.get(pk=id)
+#     return render(request, 'articles/edit.html', {'article':article})
 
 def update(request, id):
     article = Article.objects.get(pk=id)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-    return redirect(f'/articles/{article.id}/')
+    if request.method == "POST":
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect('articles:detail', article.id)
+    else:
+        return render(request, 'articles/edit.html', {'article':article})
